api/requirements.txt
Django==2.0
djangorestframework==3.7.3
pytz==2017.3
gunicorn==19.7.1
psycopg2==2.7.1